refactor: split library schema modules and add migration test
