+++ src/sample_sources/library/tests.rs
@@ -1,11 +1,102 @@
 use super::*;
 use rusqlite::OptionalExtension;
+use std::collections::HashSet;
 use std::fs;
 use tempfile::tempdir;
 
 fn with_config_home<T>(dir: &Path, f: impl FnOnce() -> T) -> T {
     let _guard = crate::app_dirs::ConfigBaseGuard::set(dir.to_path_buf());
     f()
+}
+
+const LEGACY_SCHEMA_SQL: &str = r#"
+CREATE TABLE metadata (
+    key TEXT PRIMARY KEY,
+    value TEXT NOT NULL
+);
+CREATE TABLE sources (
+    id TEXT PRIMARY KEY,
+    root TEXT NOT NULL,
+    sort_order INTEGER NOT NULL
+);
+CREATE TABLE collections (
+    id TEXT PRIMARY KEY,
+    name TEXT NOT NULL,
+    export_path TEXT,
+    sort_order INTEGER NOT NULL
+);
+CREATE TABLE collection_members (
+    collection_id TEXT NOT NULL,
+    source_id TEXT NOT NULL,
+    relative_path TEXT NOT NULL,
+    sort_order INTEGER NOT NULL,
+    PRIMARY KEY (collection_id, source_id, relative_path),
+    FOREIGN KEY(collection_id) REFERENCES collections(id) ON DELETE CASCADE
+);
+CREATE TABLE analysis_jobs (
+    id INTEGER PRIMARY KEY AUTOINCREMENT,
+    sample_id TEXT NOT NULL,
+    job_type TEXT NOT NULL,
+    status TEXT NOT NULL,
+    attempts INTEGER NOT NULL DEFAULT 0,
+    created_at INTEGER NOT NULL,
+    last_error TEXT,
+    UNIQUE(sample_id, job_type)
+);
+CREATE INDEX idx_analysis_jobs_status_created_id
+    ON analysis_jobs (status, created_at, id);
+CREATE INDEX idx_analysis_jobs_status_sample_id
+    ON analysis_jobs (status, sample_id);
+CREATE TABLE samples (
+    sample_id TEXT PRIMARY KEY,
+    content_hash TEXT NOT NULL,
+    size INTEGER NOT NULL,
+    mtime_ns INTEGER NOT NULL
+);
+CREATE TABLE analysis_features (
+    sample_id TEXT PRIMARY KEY,
+    content_hash TEXT NOT NULL,
+    features BLOB
+);
+CREATE TABLE embeddings (
+    sample_id TEXT PRIMARY KEY,
+    model_id TEXT NOT NULL,
+    dim INTEGER NOT NULL,
+    vec_blob BLOB NOT NULL
+) WITHOUT ROWID;
+"#;
+
+fn create_legacy_schema(conn: &Connection) {
+    conn.execute_batch(LEGACY_SCHEMA_SQL).unwrap();
+}
+
+fn table_columns(conn: &Connection, table: &str) -> HashSet<String> {
+    let mut stmt = conn
+        .prepare(&format!("PRAGMA table_info({})", table))
+        .unwrap();
+    stmt.query_map([], |row| row.get::<_, String>(1))
+        .unwrap()
+        .filter_map(Result::ok)
+        .collect()
+}
+
+fn assert_table_exists(conn: &Connection, table: &str) {
+    let exists: Option<String> = conn
+        .query_row(
+            "SELECT name FROM sqlite_master WHERE type='table' AND name = ?1",
+            [table],
+            |row| row.get(0),
+        )
+        .optional()
+        .unwrap();
+    assert_eq!(exists.as_deref(), Some(table));
+}
+
+fn assert_has_columns(conn: &Connection, table: &str, columns: &[&str]) {
+    let present = table_columns(conn, table);
+    for column in columns {
+        assert!(present.contains(*column), "expected {}.{}", table, column);
+    }
 }
 
 #[test]
@@ -177,5 +268,40 @@ fn reuses_known_source_id_for_same_root() {
             .unwrap()
             .expect("expected mapping");
         assert_eq!(reused.as_str(), id.as_str());
+    });
+}
+
+#[test]
+fn migrates_legacy_schema_to_latest() {
+    let temp = tempdir().unwrap();
+    with_config_home(temp.path(), || {
+        let db_path = database_path().unwrap();
+        if let Some(parent) = db_path.parent() {
+            fs::create_dir_all(parent).unwrap();
+        }
+        let conn = Connection::open(&db_path).unwrap();
+        create_legacy_schema(&conn);
+        drop(conn);
+
+        let _ = load().unwrap();
+        let conn = Connection::open(database_path().unwrap()).unwrap();
+
+        assert_has_columns(&conn, "analysis_jobs", &["content_hash"]);
+        assert_has_columns(
+            &conn,
+            "samples",
+            &["duration_seconds", "sr_used", "analysis_version"],
+        );
+        assert_has_columns(&conn, "collection_members", &["clip_root"]);
+
+        let embedding_columns = table_columns(&conn, "embeddings");
+        for column in ["vec", "dtype", "l2_normed", "created_at"] {
+            assert!(embedding_columns.contains(column));
+        }
+        assert!(!embedding_columns.contains("vec_blob"));
+
+        for table in ["features", "layout_umap", "hdbscan_clusters", "ann_index_meta"] {
+            assert_table_exists(&conn, table);
+        }
     });
 }
